move  sun utilities to new file
